requirements.txt
google-genai>=1.0.0
pinecone>=5.0.0
pypdf>=5.0.0
tqdm>=4.66.0
